Review comments: Move invesalius_pubsub dir to invesalius/pubsub
