+++ app.py
@@ -53,7 +53,7 @@ try:
 except ImportError:
     from wx import SplashScreen
 
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 #import wx.lib.agw.advancedsplash as agw
 #if sys.platform.startswith('linux'):
