+++ invesalius/data/editor.py
@@ -18,7 +18,7 @@
 #--------------------------------------------------------------------------
 
 import math
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 import vtk
 
 AXIAL = 2
