+++ invesalius/data/imagedata_utils.py
@@ -27,7 +27,7 @@ import imageio
 import numpy
 import numpy as np
 import vtk
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 from scipy.ndimage import shift, zoom
 from vtk.util import numpy_support
