+++ invesalius/data/slice_.py
@@ -22,7 +22,7 @@ import tempfile
 import numpy as np
 import vtk
 from scipy import ndimage
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 import invesalius.constants as const
 import invesalius.data.converters as converters
