+++ invesalius/data/styles_3d.py
@@ -23,7 +23,7 @@ import wx
 import invesalius.constants as const
 import invesalius.project as prj
 
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 
 PROP_MEASURE = 0.8
