+++ invesalius/data/trigger.py
@@ -21,7 +21,7 @@ import threading
 from time import sleep
 
 import wx
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 
 class Trigger(threading.Thread):
