+++ invesalius/gui/data_notebook.py
@@ -34,7 +34,7 @@ except ImportError:
     import wx.lib.flatnotebook as fnb
 
 import wx.lib.platebtn as pbtn
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 import invesalius.constants as const
 import invesalius.data.slice_ as slice_
