+++ invesalius/gui/frame.py
@@ -43,7 +43,7 @@ import wx.lib.popupctl as pc
 from invesalius import inv_paths
 from invesalius.gui import project_properties
 from wx.lib.agw.aui.auibar import AUI_TB_PLAIN_BACKGROUND, AuiToolBar
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 try:
     from wx.adv import TaskBarIcon as wx_TaskBarIcon
