+++ invesalius/gui/import_network_panel.py
@@ -19,7 +19,7 @@
 import wx
 import sys
 import wx.gizmos as gizmos
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 import wx.lib.splitter as spl
 
 import invesalius.constants as const
