+++ invesalius/gui/import_panel.py
@@ -18,7 +18,7 @@
 #--------------------------------------------------------------------------
 import wx
 import wx.gizmos as gizmos
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 import wx.lib.splitter as spl
 
 import invesalius.constants as const
