+++ invesalius/gui/task_navigator.py
@@ -42,7 +42,7 @@ except ImportError:
 
 import wx.lib.colourselect as csel
 import wx.lib.masked.numctrl
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 from time import sleep
 
 import invesalius.constants as const
