+++ invesalius/gui/widgets/clut_raycasting.py
@@ -24,7 +24,7 @@ import sys
 
 import numpy
 import wx
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 import invesalius.gui.dialogs as dialog
 import invesalius.constants as const
