+++ invesalius/plugins.py
@@ -24,7 +24,7 @@ import pathlib
 import sys
 from itertools import chain
 
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 import invesalius.constants as consts
 from invesalius import inv_paths
