+++ invesalius/project.py
@@ -30,7 +30,7 @@ import numpy as np
 import vtk
 import wx
 
-from invesalius_pubsub import pub as Publisher
+from invesalius.pubsub import pub as Publisher
 
 import invesalius.constants as const
 import invesalius.data.polydata_utils as pu
